LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssSkew.tentative-expected.txt


FAIL Constructing a CSSSkew with a keyword throws a TypeError assert_throws_js: function "() => new CSSSkew(value, CSS.deg(0))" threw object "ReferenceError: Can't find variable: CSSSkew" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkew with a double throws a TypeError assert_throws_js: function "() => new CSSSkew(value, CSS.deg(0))" threw object "ReferenceError: Can't find variable: CSSSkew" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkew with a unitless zero throws a TypeError assert_throws_js: function "() => new CSSSkew(value, CSS.deg(0))" threw object "ReferenceError: Can't find variable: CSSSkew" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkew with a string angle throws a TypeError assert_throws_js: function "() => new CSSSkew(value, CSS.deg(0))" threw object "ReferenceError: Can't find variable: CSSSkew" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkew with a number CSSUnitValue throws a TypeError assert_throws_js: function "() => new CSSSkew(value, CSS.deg(0))" threw object "ReferenceError: Can't find variable: CSSSkew" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkew with a time dimension CSSUnitValue throws a TypeError assert_throws_js: function "() => new CSSSkew(value, CSS.deg(0))" threw object "ReferenceError: Can't find variable: CSSSkew" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkew with a CSSMathValue of length type throws a TypeError assert_throws_js: function "() => new CSSSkew(value, CSS.deg(0))" threw object "ReferenceError: Can't find variable: CSSSkew" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSSkew.ax with a keyword throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ax with a double throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ax with a unitless zero throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ax with a string angle throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ax with a number CSSUnitValue throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ax with a time dimension CSSUnitValue throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ax with a CSSMathValue of length type throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ay with a keyword throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ay with a double throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ay with a unitless zero throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ay with a string angle throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ay with a number CSSUnitValue throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ay with a time dimension CSSUnitValue throws a TypeError Can't find variable: CSSSkew
FAIL Updating CSSSkew.ay with a CSSMathValue of length type throws a TypeError Can't find variable: CSSSkew
FAIL CSSSkew can be constructed from an angle CSSUnitValue and an angle CSSUnitValue Can't find variable: CSSSkew
FAIL CSSSkew can be constructed from an angle CSSUnitValue and a CSSMathValue of angle type Can't find variable: CSSSkew
FAIL CSSSkew can be constructed from a CSSMathValue of angle type and an angle CSSUnitValue Can't find variable: CSSSkew
FAIL CSSSkew can be constructed from a CSSMathValue of angle type and a CSSMathValue of angle type Can't find variable: CSSSkew
FAIL CSSSkew.ax can be updated to an angle CSSUnitValue Can't find variable: CSSSkew
FAIL CSSSkew.ax can be updated to a CSSMathValue of angle type Can't find variable: CSSSkew
FAIL CSSSkew.ay can be updated to an angle CSSUnitValue Can't find variable: CSSSkew
FAIL CSSSkew.ay can be updated to a CSSMathValue of angle type Can't find variable: CSSSkew
FAIL Modifying CSSSkew.is2D is a no-op Can't find variable: CSSSkew
